requirements.txt
asgiref==3.6.0
beautifulsoup4==4.10.0
blis==0.7.9
catalogue==2.0.8
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
Django==4.1.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
fonttools==4.39.0
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.2
matplotlib==3.7.1
murmurhash==1.0.9
mysql==0.0.3
mysql-connector-python==8.0.32
mysqlclient==2.1.1
nltk==3.6.5
numpy==1.22.4
packaging==23.0
pandas==1.3.4
pathy==0.10.1
Pillow==9.4.0
preshed==3.0.8
protobuf==3.20.3
pycryptodome==3.16.0
pydantic==1.8.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
regex==2022.10.31
requests==2.26.0
scikit-learn==1.2.1
scipy==1.7.1
six==1.16.0
smart-open==6.3.0
soupsieve==2.4
spacy==3.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.3
srsly==2.4.6
thinc==8.0.17
threadpoolctl==3.1.0
tqdm==4.65.0
typer==0.4.2
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
wasabi==0.10.1
wincertstore==0.2
wordcloud==1.8.2.2
zipp==3.15.0
